refactor: separate versioning
